requirements.txt
pycrypto==2.6.1
base58==2.0.1
ecdsa==0.16.1
requests==2.25.1
requests-oauthlib==1.3.0
cryptography==3.4.4
python-nmap==0.7.1
rich==9.5.1
pysha3==1.0.2
sha3==0.2.1
pycryptodomex==3.12.0
fixedint==0.1.6
eth-rlp==0.2.1
rlp==2.0.1
leveldb==0.201
Gemtography==0.0.1
